Add entity collision with blocks.
