+++ src/MiNET/MiNET/Entities/Mob.cs
@@ -17,49 +17,5 @@ namespace MiNET.Entities
 		{
 		}
 
-		public override void OnTick()
-		{
-			base.OnTick();
-
-			if (Velocity.Distance > 0)
-			{
-				PlayerLocation oldPosition = (PlayerLocation) KnownPosition.Clone();
-				bool onGroundBefore = IsOnGround(KnownPosition);
-
-				KnownPosition.X += (float) Velocity.X;
-				KnownPosition.Y += (float) Velocity.Y;
-				KnownPosition.Z += (float) Velocity.Z;
-
-				bool onGround = IsOnGround(KnownPosition);
-				if (!onGroundBefore && onGround)
-				{
-					KnownPosition.Y = (float) Math.Floor(oldPosition.Y);
-					Velocity = Vector3.Zero;
-				}
-				else
-				{
-					Velocity *= (1.0 - Drag);
-					if (!onGround)
-					{
-						Velocity -= new Vector3(0, Gravity, 0);
-					}
-				}
-				LastUpdatedTime = DateTime.UtcNow;
-			}
-			else if (Velocity != Vector3.Zero)
-			{
-				Velocity = Vector3.Zero;
-				LastUpdatedTime = DateTime.UtcNow;
-			}
-		}
-
-		private bool IsOnGround(PlayerLocation position)
-		{
-			PlayerLocation pos = (PlayerLocation) position.Clone();
-			pos.Y -= 0.1f;
-			Block block = Level.GetBlock(new BlockCoordinates(pos));
-
-			return block.Id != 0; // Should probably test for solid
-		}
 	}
 }
